Extract and test export table creation logic
I had to disable formating date fields as date field is not a number anymore.
