+++ packer/jobs/patient_diagnosis_biosource_biomaterial_export.py
@@ -1,15 +1,8 @@
 import logging
-import datetime as dt
 from zipfile import ZipFile
-import pandas as pd
-import numpy as np
 import requests
 from celery.exceptions import Ignore
-
-try:
-    from transmart.api.v2.data_structures import ObservationSet
-except ImportError as e:
-    logging.warning(f'Import errors for {__file__!r}: {str(e)}')
+from ..table_transformations.patient_diagnosis_biosource_biomaterial import to_patient_diagnosis_biosource_biomaterial_dataframe
 
 from packer.file_handling import FSHandler
 from packer.task_status import Status
@@ -18,8 +11,6 @@ from ..tasks import BaseDataTask, app
 
 logger = logging.getLogger(__name__)
 SEP = '\t'
-DATE_FORMAT = '%d-%m-%Y'
-IDENTIFYING_COLUMN_LIST = ['Patient ID', 'PMC Diagnosis ID', 'PMC Biosource ID', 'PMC Biomaterial ID']
 
 
 @app.task(bind=True, base=BaseDataTask)
@@ -38,7 +29,7 @@ def patient_diagnosis_biosource_biomaterial_export(self: BaseDataTask, constrain
     r = requests.post(url=handle,
                       json={'type': 'clinical', 'constraint': constraint},
                       headers={
-                             'Authorization': self.request.get('Authorization')
+                          'Authorization': self.request.get('Authorization')
                       })
 
     if r.status_code == 401:
@@ -53,140 +44,11 @@ def patient_diagnosis_biosource_biomaterial_export(self: BaseDataTask, constrain
         raise Ignore()
 
     self.update_status(Status.RUNNING, 'Observations gotten, transforming.')
-    obs = ObservationSet(r.json()).dataframe
-    reformatted_obs = reformat_export(obs)
+    reformatted_obs = to_patient_diagnosis_biosource_biomaterial_dataframe(r.json())
 
     self.update_status(Status.RUNNING, 'Writing export to disk.')
     with FSHandler(self.task_id).writer as writer:
         with ZipFile(writer, 'w') as data_zip:
-            data_zip.writestr('data.tsv', reformatted_obs.to_csv(encoding='utf-8', sep='\t', index=False))
+            data_zip.writestr('data.tsv', reformatted_obs.to_csv(encoding='utf-8', sep=SEP, index=False))
 
     logger.info(f'Stored to disk: {self.task_id}')
-
-
-def reformat_export(obs):
-    if obs.empty:
-        logger.warn('Retrieved hypercube is empty! Exporting empty result.')
-        return obs
-
-    # order rows by concept_paths:
-    # 1)Patient -> 2)Diagnosis -> 3)Biosource -> 4)Biomaterial -> 5)Studies
-    obs.rename(columns={'patient.trial': 'Patient ID'}, inplace=True)
-    obs.sort_values(by=['concept.conceptPath'], inplace=True)
-    concept_order = obs['concept.name'].unique().tolist()
-    id_columns = get_identifying_columns(obs)
-
-    # reformat columns: rename, drop, merge
-    logger.info('Reformatting columns...')
-    obs = reformat_columns(obs, id_columns)
-    # transform concept rows to column headers
-    obs_pivot = concepts_row_to_columns(obs, concept_order)
-
-    # propagate data to lower levels and display only rows that represent the lowest level
-    obs_pivot = rebuild_rows(obs_pivot, id_columns)
-    obs_pivot.reset_index(drop=True, inplace=True)
-
-    # fill NaNs with empty string
-    obs_pivot.fillna('', inplace=True)
-
-    # update date fields with format
-    update_date_fields(obs_pivot, obs)
-    return obs_pivot
-
-
-def drop_higher_level(lowest_level_column, data):
-    lowest_level = data[data[lowest_level_column].notnull()]
-    column_index = data.columns.get_loc(lowest_level_column)
-    for index, row in lowest_level.iterrows():
-        if column_index == 3:
-            conditions = ((data[data.columns[0]] == row[0]) &
-                          (data[data.columns[1]] == row[1]) &
-                          (data[data.columns[2]] == row[2]) &
-                          (data[data.columns[3]].isnull()))
-        elif column_index == 2:
-            conditions = ((data[data.columns[0]] == row[0]) &
-                          (data[data.columns[1]] == row[1]) &
-                          (data[data.columns[2]].isnull()) &
-                          (data[data.columns[3]].isnull()))
-        else:
-            conditions = ((data[data.columns[0]] == row[0]) &
-                          (data[data.columns[1]].isnull()) &
-                          (data[data.columns[2]].isnull()) &
-                          (data[data.columns[3]].isnull()))
-        data.drop(data[conditions].index, inplace=True)
-
-
-def limit_rows_to_lowest_level(data, id_columns):
-    logger.info('Removing redundant rows...')
-    lowest_level_col_id = len(id_columns)-1
-    for idx, column in enumerate(reversed(id_columns)):
-        if idx == lowest_level_col_id:
-            break
-        drop_higher_level(column, data)
-
-
-def rebuild_rows(data, id_columns):
-    # sort rows by identifying columns
-    data.sort_values(id_columns, na_position='first', inplace=True)
-    grouped_data = data.groupby(id_columns[0])
-    # propagate data to lower levels
-    ffill_data = grouped_data.ffill()
-    # limit rows to the lowest level
-    limit_rows_to_lowest_level(ffill_data, id_columns)
-    return ffill_data
-
-
-def get_date_string(timestamp, string_format=DATE_FORMAT):
-    if pd.notnull(timestamp) and timestamp is not None and timestamp != '':
-        return dt.datetime.utcfromtimestamp(float(timestamp) / 1000).strftime(string_format)
-    else:
-        return timestamp
-
-
-def update_date_fields(data, obs):
-    for column in data.columns:
-        if column.upper().startswith('DATE'):
-            obs[column] = data[column].apply(get_date_string)
-
-
-def reformat_columns(obs, id_columns):
-    # rename columns and set indexes
-    obs.reset_index(inplace=True)
-    headers = np.append(id_columns, 'concept.name')
-
-    # prepare 'value' column
-    if {'stringValue', 'numericValue'}.issubset(obs.columns):
-        # merge stringValue and numericValue into value column
-        obs['value'] = obs['stringValue'].fillna(obs['numericValue'])
-        obs.drop(['stringValue', 'numericValue'], axis=1, inplace=True)
-    elif 'stringValue' in obs:
-        obs.rename(columns={"stringValue": "value"}, inplace=True)
-    elif 'numericValue' in obs:
-        obs.rename(columns={"numericValue": "value"}, inplace=True)
-    else:
-        obs['value'] = ""
-    obs = obs.set_index(list(headers), append=True)[['value']]
-
-    return obs
-
-
-def concepts_row_to_columns(obs, concept_order):
-    # use unstack to move the last level of the index to column names
-    obs_pivot = obs.unstack(level=-1)
-    # update column names by dropping value level
-    obs_pivot.columns = obs_pivot.columns.droplevel(level=0)
-    # fix the order of concept columns
-    obs_pivot = obs_pivot[concept_order]
-    # fix indexes
-    obs_pivot.reset_index(inplace=True)
-    obs_pivot.drop(obs_pivot.columns[[0]], axis=1, inplace=True)
-    return obs_pivot
-
-
-def get_identifying_columns(obs):
-    columns = []
-    for id_column in IDENTIFYING_COLUMN_LIST:
-        if id_column in obs:
-            columns.append(id_column)
-    logger.info(columns)
-    return columns
