+++ packer/table_transformations/__init__.py
@@ -1,0 +1,5 @@
+from .patient_diagnosis_biosource_biomaterial import to_patient_diagnosis_biosource_biomaterial_dataframe
+
+registry = {
+    'reformat_export': to_patient_diagnosis_biosource_biomaterial_dataframe,
+}
